smartplay/requirements.txt
transformers==4.41.2
bitsandbytes==0.42.0
accelerate==0.31.0
gym==0.21.0
numpy==2.1.3
pandas==2.2.2
tqdm==4.67.0
wandb==0.18.7
openai==1.93.0
python-dotenv
pyyaml
sentencepiece==0.2.0
pytest
ollama==0.3.3
huggingface-hub==0.26.2
matplotlib==3.9.3
seaborn==0.13.2
scikit-learn==1.6.1
Pillow>=9.4.0
imageio==2.34.0